Change TIFF SubIFD offset size to uint64_t in TPTImage.h. Fixes https://github.com/ruven/iipsrv/issues/281

+++ src/TPTImage.h
@@ -39,7 +39,7 @@ class TPTImage : public IIPImage {
   TIFF *tiff;
 
   /// List of SubIFD sub-resolutions
-  std::vector<uint32_t> subifds;
+  std::vector<uint64_t> subifds;
 
   /// To which IFD do these SubIFDs belong
   tdir_t subifd_ifd;
